Completion of refactor and validation that this still works.

+++ load-test/checkout_simulator.sh
@@ -1,14 +1,14 @@
 #!/bin/bash
 #
-# This script supports load testing against Magento 1.8 CE and 1.13 EE installations
-# It forks subprocesses that perform the same HTTP calls as users with browsers
-# Test outputs and analyses should come from third parties; this is just a runner
+# This script supports load testing against Magento 1.8 CE and 1.13 EE installations;
+# it forks subprocesses that perform the same HTTP calls as users with browsers.
+# Test outputs and analyses should come from third parties; this is just a runner.
 #
-# See https://github.com/copious/cop_magento_tests for additional information
+# See https://github.com/copious/cop_magento_tests for additional information.
 #
-# Typical outcome: X orders per hour for environment Y
+# Typical outcome: X orders per hour for environment Y.
 #
-# Tested with Ubuntu 12.04 LTS, GNU bash 4.2.25, and curl 7.22.0 against Magento 1.13.1 EE
+# Tested with Ubuntu 12.04 LTS, GNU bash 4.2.25, and curl 7.22.0 against Magento 1.13.1 EE.
 #
 # This script will probably not work with previous Magento versions as the
 # form_key parameter is expected to be present.
@@ -16,6 +16,8 @@
 # CONFIGURATION OPTIONS
 ######
 
+# Debug mode is noisy
+DEBUG=false
 # Protocol: http or https
 PROTO=http
 # Host: domain name, ip address; ports optional
@@ -31,6 +33,8 @@ INTRA_USER_SLEEP=0.75
 CURL_CONFIG_FILE=checkout_config
 # a file or full path to log outcomes of subprocesses
 LOG_FILE=checkout_simulator.log
+# a file or full path to store ids of subprocesses
+PID_FILE=checkout_simulator.pid
 # how many of each product to add to the cart?
 QTY=1
 # URL parameters for the /checkout/onepage/saveMethod/ step
@@ -41,23 +45,34 @@ SHIPPING_METHOD="shipping_method=flatrate_flatrate"
 # NOTE: use test cards https://www.paypalobjects.com/en_US/vhelp/paypalmanager_help/credit_card_numbers.htm
 PAYMENT_METHOD="payment[method]=checkmo"
 
+echo -e "\nCheckout simulator\nhost: ${HOST}\n"
+
 if $FASTER
 then
-  CONCURRENCIES=(1 2 5)
-  CONCURRENCY_DURATION=120
+  CONCURRENCIES=(1 2 4)
+  CONCURRENCY_DURATION=300
 else
   # Bash array of counts of threads to run
-  CONCURRENCIES=(3 6 12 25 30 35 40 50)
+  CONCURRENCIES=(10 20 40 60 80 100 120 140)
   # how long to run each level of concurrency, in seconds
-  CONCURRENCY_DURATION=1200
+  CONCURRENCY_DURATION=1800
+fi
+
+if $DEBUG
+then
+  FASTER=true
+  CONCURRENCIES=(1)
+  CONCURRENCY_DURATION=300
+  echo -e "DEBUG MODE\n"
 fi
 
 if $FASTER
 then
-  PAGE_SLEEP=1
-  CHECKOUT_SLEEP=5
-  AJAX_SLEEP=$(( $PAGE_SLEEP * 2 ))
+  PAGE_SLEEP=0.25
+  CHECKOUT_SLEEP=2
+  AJAX_SLEEP=`bc -l <<< "$PAGE_SLEEP * 2"`
   RETRY_COUNT=2
+  echo -e "FAST MODE\n"
 else
   # prime numbers can spread out the load
   # seconds between each page load
@@ -83,13 +98,22 @@ POST_CONTENT_TYPE="Content-type:application/x-www-form-urlencoded; charset=UTF-8
 
 # This function is called when this script is told to quit
 function clean_up {
-  echo -e "\n\nPausing to clean up session files…"
-  # Kill all sleep operations from the current user
-  killall -u `whoami` sleep
-  # This *should* wait for all subprocesses to stop
-  wait
-  rm -f *.cookie
-  rm -f *.tmp
+  # check if this process has forked children
+  job_count=`jobs -p | wc -l`
+  if [ $job_count -gt 0 ]
+  then
+    printf "\n\nCleaning up session files…"
+    # Kill all sub processes, if possible
+    if [ -e $PID_FILE ]
+    then
+      cat $PID_FILE | xargs kill > $LOG_FILE 2>&1
+    fi
+    # This *should* wait for all subprocesses to stop
+    wait
+    rm -f *.cookie
+    rm -f *.tmp
+    printf " done.\n\n"
+  fi
   exit
 }
 
@@ -102,21 +126,30 @@ timed_shop() {
   # arg 1 is how long to run for
   local MY_DURATION=$1
   # find end time
-  local MY_SECONDS_TARGET=$(($SECONDS + $MY_DURATION))
+  local MY_SECONDS_TARGET=$(( $SECONDS + $MY_DURATION ))
   # random temp/state files
   local RAND=`LC_CTYPE=C tr -dc A-Za-z0-9 < /dev/urandom | head -c 32 | xargs`
   local COOKIE="${RAND}.cookie"
   local TEMPFILE="${RAND}.tmp"
+  local MY_CHECKOUT_SUCCESS=0
+  local MY_CHECKOUT_ATTEMPTS=0
+  local MY_ERROR_COUNT=0
 
   # Perform the actions for x many seconds
   # Note that if this loop is really long the rate of orders will be skewed
   while [ $SECONDS -le $MY_SECONDS_TARGET ]
   do
     local START=$SECONDS
-    echo -e "------\nUser ${RAND} (${2} of ${3})"
-    echo "  iteration ends at: ${MY_SECONDS_TARGET}"
-    echo "  time elapsed: ${SECONDS}"
-    echo -e "  time to shop: ${MY_DURATION}\n"
+    if $DEBUG
+    then
+      echo -e "------\nUser:${RAND}-${2}/${3}"
+      echo "  iteration ends at: ${MY_SECONDS_TARGET}"
+      echo "  time elapsed: ${SECONDS}"
+      echo "  time to shop: ${MY_DURATION}"
+      echo "  checkouts attempted: ${MY_CHECKOUT_ATTEMPTS}"
+      echo "  checkouts successful: ${MY_CHECKOUT_SUCCESS}"
+      echo "  error count: ${MY_ERROR_COUNT}"
+    fi
 
     # Naively assuming these are writable
     touch $COOKIE
@@ -139,19 +172,30 @@ timed_shop() {
     curl -K $CURL_CONFIG_FILE -b $COOKIE -H $HTML_ACCEPT -c $COOKIE ${PROTO}://${HOST}/natural-science/science-history/simple-001196-product.html > $TEMPFILE 2>&1
     # Find the form_key! (feature introduced with 1.13)
     FORM_KEY=`cat $TEMPFILE | grep -F -m 1 'form_key' | sed 's/.*\/form_key\/\([^\/]*\).*/\1/'`
-    echo -e "\nFORM_KEY: ${FORM_KEY}" >> $LOG_FILE
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} FORM_KEY: ${FORM_KEY}"
+    fi
 
     curl -K $CURL_CONFIG_FILE -b $COOKIE -H $HTML_ACCEPT -c $COOKIE ${PROTO}://${HOST}/natural-science/science-history/simple-001196-product.html > $TEMPFILE 2>&1
     # find the cart URL
     CART_ACTION=$(find_cart_url $TEMPFILE)
-
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} CART_ACTION: product 1 ${CART_ACTION}"
+    fi
     sleep $PAGE_SLEEP
     # add to cart
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} ADD: product 1 to cart"
+    fi
     curl -K $CURL_CONFIG_FILE -b $COOKIE -c $COOKIE --retry $RETRY_COUNT -H $AJAX_ACCEPT -d "form_key=${FORM_KEY}&product=1196&related_product=&qty=${QTY}" ${CART_ACTION} > /dev/null 2>&1
     # cart page
     load_url $COOKIE "${CART_URL}" > /dev/null 2>&1
     # home page
     load_url $COOKIE "${PROTO}://${HOST}/" > /dev/null 2>&1
+    sleep $PAGE_SLEEP
     
     ## PRODUCT 2
     ######
@@ -162,8 +206,17 @@ timed_shop() {
     load_url $COOKIE "${PROTO}://${HOST}/philosophy/eastern-philosophy/simple-002327-product.html" > $TEMPFILE 2>&1
     # find the cart URL
     CART_ACTION=$(find_cart_url $TEMPFILE)
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} CART_ACTION: product 2 ${CART_ACTION}"
+    fi
+    sleep $PAGE_SLEEP
 
     # add to cart
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} ADD: product 2 to cart"
+    fi
     curl -K $CURL_CONFIG_FILE -b $COOKIE -c $COOKIE --retry $RETRY_COUNT -H $AJAX_ACCEPT -d "form_key=${FORM_KEY}&product=2327&related_product=&qty=${QTY}" ${CART_ACTION} > /dev/null 2>&1
     # cart page
     load_url $COOKIE "${CART_URL}" > /dev/null 2>&1
@@ -178,8 +231,17 @@ timed_shop() {
     load_url $COOKIE "${PROTO}://${HOST}/natural-science/environmental-science/simple-001239-product.html" > $TEMPFILE 2>&1
     # find the cart URL
     CART_ACTION=$(find_cart_url $TEMPFILE)
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} CART_ACTION: product 3 ${CART_ACTION}"
+    fi
+    sleep $PAGE_SLEEP
     
     # add to cart
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} ADD: product 3 to cart"
+    fi
     curl -K $CURL_CONFIG_FILE -b $COOKIE -c $COOKIE --retry $RETRY_COUNT -H $AJAX_ACCEPT -d "form_key=${FORM_KEY}&product=1239&related_product=&qty=${QTY}" ${CART_ACTION} > /dev/null 2>&1
     # cart page
     load_url $COOKIE "${CART_URL}" > /dev/null 2>&1
@@ -194,8 +256,17 @@ timed_shop() {
     load_url $COOKIE "${PROTO}://${HOST}/computer/computer-engineering/simple-001139-product.html" > $TEMPFILE 2>&1
     # find the cart URL
     CART_ACTION=$(find_cart_url $TEMPFILE)
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} CART_ACTION: product 4 ${CART_ACTION}"
+    fi
+    sleep $PAGE_SLEEP
     
     # add to cart
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} ADD: product 4 to cart"
+    fi
     curl -K $CURL_CONFIG_FILE -b $COOKIE -c $COOKIE --retry $RETRY_COUNT -H $AJAX_ACCEPT -d "form_key=${FORM_KEY}&product=1139&related_product=&qty=${QTY}" ${CART_ACTION} > /dev/null 2>&1
     # cart page
     load_url $COOKIE "${CART_URL}" > /dev/null 2>&1
@@ -210,8 +281,17 @@ timed_shop() {
     load_url $COOKIE "${PROTO}://${HOST}/natural-science/environmental-science/simple-001522-product.html" > $TEMPFILE 2>&1
     # find the cart URL
     CART_ACTION=$(find_cart_url $TEMPFILE)
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} CART_ACTION: product 5 ${CART_ACTION}"
+    fi
+    sleep $PAGE_SLEEP
     
     # add to cart
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} ADD: product 5 to cart"
+    fi
     curl -K $CURL_CONFIG_FILE -b $COOKIE -c $COOKIE --retry $RETRY_COUNT -H $AJAX_ACCEPT -d "form_key=${FORM_KEY}&product=1522&related_product=&qty=${QTY}" ${CART_ACTION} > /dev/null 2>&1
     # cart page
     load_url $COOKIE "${CART_URL}" > /dev/null 2>&1
@@ -220,6 +300,10 @@ timed_shop() {
     ## CART, SHIPPING, TAX
     ######
 
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} LOADING: cart"
+    fi
     load_url $COOKIE "${CART_URL}" > /dev/null 2>&1
     sleep $CHECKOUT_SLEEP
 
@@ -228,27 +312,50 @@ timed_shop() {
     ## CHECKOUT
     ######
     
-    echo -e "\nUser ${RAND} checking out"
+    MY_CHECKOUT_ATTEMPTS=$(( $MY_CHECKOUT_ATTEMPTS + 1 ))
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} LOADING: checkout"
+    fi
 
     load_url $COOKIE "${PROTO}://${HOST}/checkout/onepage/" > $TEMPFILE 2>&1
     # Find the address ID!
+    # NOTE: not implemented for guest checkout
     ADDRESS_ID=`cat $TEMPFILE | grep 'address_id' | sed 's/[^0-9]*//g'`
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} ADDRESS_ID: ${ADDRESS_ID}"
+      echo "User:${RAND}-${2}/${3} SAVING: ${CHECKOUT_METHOD}"
+    fi
 
     curl -K $CURL_CONFIG_FILE -b $COOKIE -c $COOKIE --retry $RETRY_COUNT -H $AJAX_ACCEPT -H $CHECKOUT_REFERER -H $AJAX_REQUESTED_WITH -d $CHECKOUT_METHOD ${PROTO}://${HOST}/checkout/onepage/saveMethod/ > /dev/null 2>&1
 
     # NOTE: this is where users could sign in or sign up
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} LOADING: billing fields"
+    fi
 
     load_ajax_checkout $COOKIE"${PROTO}://${HOST}/checkout/onepage/progress/?prevStep=billing" > /dev/null 2>&1
+    sleep $CHECKOUT_SLEEP
 
     # Save billing
-    sleep $CHECKOUT_SLEEP
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} SAVING: billing info"
+    fi
     curl -K $CURL_CONFIG_FILE -b $COOKIE -c $COOKIE --retry $RETRY_COUNT -H $AJAX_ACCEPT -H $CHECKOUT_REFERER -H $AJAX_REQUESTED_WITH -d "billing[address_id]=&billing[firstname]=Test&billing[lastname]=Account&billing[company]=&billing[email]=test${RAND}@example.org&billing[street][]=411 SW 6th Ave&billing[street][]=&billing[city]=Portland&billing[region_id]=49&billing[region]=&billing[postcode]=97204&billing[country_id]=US&billing[telephone]=5035551212&billing[fax]=&billing[customer_password]=&billing[confirm_password]=&billing[save_in_address_book]=1&billing[use_for_shipping]=1" ${PROTO}://${HOST}/checkout/onepage/saveBilling/ > /dev/null 2>&1
 
-    # Get additional; this tends to trigger other Magento features
+    # Get additional; this tends to trigger other Magento features like shipping estimates
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} LOADING: additional fields"
+    fi
     curl -K $CURL_CONFIG_FILE -b $COOKIE -c $COOKIE --retry $RETRY_COUNT -H $AJAX_ACCEPT -H $CHECKOUT_REFERER -H $AJAX_REQUESTED_WITH -X POST ${PROTO}://${HOST}/checkout/onepage/getAdditional/ > $TEMPFILE 2>&1
     # Find the gift options ID
     GIFT_OPTION_LINES=`grep -F 'giftoptions' $TEMPFILE`
-    echo $GIFT_OPTION_LINES > $TEMPFILE
+    # Ensure each input is separated by newline
+    echo $GIFT_OPTION_LINES | sed "s/\/> ?</\/>\n</g" > $TEMPFILE
     # NOTE: the vanilla template has one gift option for the cart and each item
     GIFT_OPTION_ID_1=`sed 's/[^0-9]*//g' $TEMPFILE | head -n 1 | tail -n 1`
     GIFT_OPTION_ID_2=`sed 's/[^0-9]*//g' $TEMPFILE | head -n 2 | tail -n 1`
@@ -256,32 +363,88 @@ timed_shop() {
     GIFT_OPTION_ID_4=`sed 's/[^0-9]*//g' $TEMPFILE | head -n 4 | tail -n 1`
     GIFT_OPTION_ID_5=`sed 's/[^0-9]*//g' $TEMPFILE | head -n 5 | tail -n 1`
     GIFT_OPTION_ID_6=`sed 's/[^0-9]*//g' $TEMPFILE | head -n 6 | tail -n 1`
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} Gift options:\n$(cat $TEMPFILE)"
+      echo "User:${RAND}-${2}/${3} GIFT_OPTION_ID_1: ${GIFT_OPTION_ID_1}"
+    fi
+    sleep $PAGE_SLEEP
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} LOADING: billing progress"
+    fi
     
     # load the progress steps
     load_ajax_checkout $COOKIE "${PROTO}://${HOST}/checkout/onepage/progress/?prevStep=billing" > /dev/null 2>&1
     sleep $PAGE_SLEEP
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} LOADING: shipping progress"
+    fi
     load_ajax_checkout $COOKIE "${PROTO}://${HOST}/checkout/onepage/progress/?prevStep=shipping" > /dev/null 2>&1
     sleep $CHECKOUT_SLEEP
 
     # Save shipping
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} SAVING: ${SHIPPING_METHOD}"
+    fi
     curl -K $CURL_CONFIG_FILE -b $COOKIE -c $COOKIE --retry $RETRY_COUNT -H $CHECKOUT_REFERER -H $AJAX_ACCEPT -H $AJAX_REQUESTED_WITH -H $POST_CONTENT_TYPE -d $SHIPPING_METHOD ${PROTO}://${HOST}/checkout/onepage/saveShippingMethod/ > /dev/null 2>&1
     sleep $PAGE_SLEEP
+
     # load the progress steps
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} LOADING: shipping progress"
+    fi
     load_ajax_checkout $COOKIE "${PROTO}://${HOST}/checkout/onepage/progress/?prevStep=shipping_method" > /dev/null 2>&1
     sleep $CHECKOUT_SLEEP
 
     # Save payment
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} SAVING: ${PAYMENT_METHOD}"
+    fi
     curl -K $CURL_CONFIG_FILE -b $COOKIE -c $COOKIE --retry $RETRY_COUNT -H $AJAX_ACCEPT -H $CHECKOUT_REFERER -H $AJAX_REQUESTED_WITH -d $PAYMENT_METHOD ${PROTO}://${HOST}/checkout/onepage/savePayment/ > /dev/null 2>&1
     sleep $PAGE_SLEEP
+
     # load the progress steps
     load_ajax_checkout $COOKIE "${PROTO}://${HOST}/checkout/onepage/progress/?prevStep=payment" > /dev/null 2>&1
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} LOADING: payment progress"
+    fi
     sleep $CHECKOUT_SLEEP
 
     # Save order
-    curl -K $CURL_CONFIG_FILE -b $COOKIE -c $COOKIE --retry $RETRY_COUNT -H $AJAX_ACCEPT -H $CHECKOUT_REFERER -H $AJAX_REQUESTED_WITH -d $PAYMENT_METHOD ${PROTO}://${HOST}/checkout/onepage/saveOrder/form_key/${FORM_KEY}/ >> $LOG_FILE 2>&1
+    echo '' > $TEMPFILE
+    curl -K $CURL_CONFIG_FILE -i -b $COOKIE -c $COOKIE --retry $RETRY_COUNT -H $AJAX_ACCEPT -H $CHECKOUT_REFERER -H $AJAX_REQUESTED_WITH -d $PAYMENT_METHOD ${PROTO}://${HOST}/checkout/onepage/saveOrder/form_key/${FORM_KEY}/ > $TEMPFILE
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} SAVING: order"
+    fi
+    success_count=$(grep -F '{"success":true,"error":false}' $TEMPFILE | wc -l)
+    if [ $success_count -gt 0 ]
+    then
+      if $DEBUG
+      then
+        echo "User:${RAND}-${2}/${3} SUCCESS!"
+      fi
+      MY_CHECKOUT_SUCCESS=$(( $MY_CHECKOUT_SUCCESS + 1 ))
+    else
+      if $DEBUG
+      then
+        echo "User:${RAND}-${2}/${3} ERROR: checkout not successful"
+      fi
+      MY_ERROR_COUNT=$(( $MY_ERROR_COUNT + 1 ))
+    fi
     sleep $CHECKOUT_SLEEP
 
     # Load reciept
+    if $DEBUG
+    then
+      echo "User:${RAND}-${2}/${3} LOADING: reciept"
+    fi
     load_url $COOKIE "${PROTO}://${HOST}/checkout/onepage/success/" > /dev/null 2>&1
     sleep $CHECKOUT_SLEEP
 
@@ -289,9 +452,14 @@ timed_shop() {
     rm -f $COOKIE
     rm -f $TEMPFILE
 
-    ELAPSED_TIME=$(($SECONDS - $START))
+    ELAPSED_TIME=$(( $SECONDS - $START ))
 
-    echo -e "\n\n------\nUser ${RAND} CHECKOUT COMPLETED (in ${ELAPSED_TIME} seconds)\n------\n\n" >> $LOG_FILE
+    if $DEBUG
+    then
+      message="\n\n------\nUser:${RAND}-${2}/${3} CHECKOUT COMPLETED (in ${ELAPSED_TIME} seconds)\n------\n\n"
+      echo -e $message >> $LOG_FILE
+      echo -e $message
+    fi
   done
 }
 
@@ -339,8 +507,6 @@ trap clean_up SIGHUP SIGINT SIGTERM EXIT
 ## RUN THE SCRIPT
 ######
 
-echo -e "\nCheckout simulator\nhost: ${HOST}\n"
-
 # concurrency notes: http://opennomad.com/content/parallelism-or-multiple-threads-bash
 for c in "${CONCURRENCIES[@]}"
 do
@@ -350,24 +516,49 @@ do
   # reset instance counter
   i=1
 
+  # create the pid file
+  touch $PID_FILE
+  if [ $? != 0 ];
+  then
+    echo "FATAL: could not write to pid file: ${PID_FILE}"
+    exit 1
+  fi
+  # clear the pids
+  echo '' > $PID_FILE
+
   START_TIME=$SECONDS
   # perform the actions for x many users
   while [ $i -le $c ]
   do
-    ELAPSED_TIME=$(($SECONDS - $START_TIME))
+    ELAPSED_TIME=$(( $SECONDS - $START_TIME ))
     # fork a process to shop
-    TIME_TO_RUN=$(($CONCURRENCY_DURATION - $ELAPSED_TIME))
+    TIME_TO_RUN=$(( $CONCURRENCY_DURATION - $ELAPSED_TIME ))
     timed_shop $TIME_TO_RUN $i $c &
+    SUBPID=$!
+    echo $SUBPID >> $PID_FILE
+    if $DEBUG
+    then
+      echo "Forking PID ${SUBPID}"
+    fi
     # increment instance counter
-    i=$(( $i + 1))
+    i=$(( $i + 1 ))
+    if $DEBUG
+    then
+      echo "Pausing for ${INTRA_USER_SLEEP} seconds"
+    fi
     # wait for the configured time period
     sleep $INTRA_USER_SLEEP
   done
+  if $DEBUG
+  then
+    echo -e "All simulated users forked; waiting for them to complete…\n"
+  fi
   # wait for the forked processes to complete
   wait
   message="END: checkout as ${c} users (`date`)"
   echo $message >> $LOG_FILE
   echo -e "${message}\n"
+  rm -f $PID_FILE
 done
 
 echo -e "Done! All concurrency scenarios have run.\n"
